SF Mismatch between package name and bundle symbolic name
