+++ modules/apps/oauth2-provider/oauth2-provider-scope-impl/src/main/java/com/liferay/oauth2/provider/scope/internal/configuration/DefaultConfigurableScopeMapperRegistrator.java
@@ -1,4 +1,4 @@
-package com.liferay.oauth2.provider.internal.configuration;
+package com.liferay.oauth2.provider.scope.internal.configuration;
 
 import com.liferay.oauth2.provider.configuration.DefaultConfigurableScopeMapperRegistratorConfiguration;
 import com.liferay.oauth2.provider.scope.impl.scopemapper.ConfigurableScopeMapper;
